+++ packages/vtable-gantt/src/event/event-manager.ts
@@ -143,6 +143,7 @@ function bindTableGroupListener(event: EventManager) {
             const taskIndex = getTaskIndexByY(e.offset.y, scene._gantt);
             const record = scene._gantt.getRecordByIndex(taskIndex);
             scene._gantt.fireListeners(GANTT_EVENT_TYPE.MOUSEENTER_TASK_BAR, {
+              federatedEvent: e,
               event: e.nativeEvent,
               index: taskIndex,
               record
@@ -156,6 +157,7 @@ function bindTableGroupListener(event: EventManager) {
             const taskIndex = getTaskIndexByY(e.offset.y, scene._gantt);
             const record = scene._gantt.getRecordByIndex(taskIndex);
             scene._gantt.fireListeners(GANTT_EVENT_TYPE.MOUSELEAVE_TASK_BAR, {
+              federatedEvent: e,
               event: e.nativeEvent,
               index: taskIndex,
               record
@@ -301,6 +303,7 @@ function bindTableGroupListener(event: EventManager) {
         const taskIndex = getTaskIndexByY(e.offset.y, gantt);
         const record = gantt.getRecordByIndex(taskIndex);
         gantt.fireListeners(GANTT_EVENT_TYPE.CLICK_TASK_BAR, {
+          federatedEvent: e,
           event: e.nativeEvent,
           index: taskIndex,
           record
@@ -329,6 +332,7 @@ function bindTableGroupListener(event: EventManager) {
         gantt.updateTaskRecord(recordTaskInfo.taskRecord, taskIndex);
         if (gantt.hasListeners(GANTT_EVENT_TYPE.CREATE_TASK_SCHEDULE)) {
           gantt.fireListeners(GANTT_EVENT_TYPE.CREATE_TASK_SCHEDULE, {
+            federatedEvent: e,
             event: e.nativeEvent,
             index: taskIndex,
             startDate: recordTaskInfo.taskRecord[gantt.parsedOptions.startDateField],
@@ -346,11 +350,24 @@ function bindTableGroupListener(event: EventManager) {
       stateManager.hideTaskBarSelectedBorder();
       scene._gantt.stateManager.selectedDenpendencyLink.link = depedencyLink;
       stateManager.showDependencyLinkSelectedLine();
+    } else if ((isClickLeftLinkPoint || isClickRightLinkPoint) && event.poniterState === 'down') {
+      if (gantt.hasListeners(GANTT_EVENT_TYPE.CLICK_DEPENDENCY_LINK_POINT)) {
+        const taskIndex = getTaskIndexByY(e.offset.y, gantt);
+        const record = gantt.getRecordByIndex(taskIndex);
+        gantt.fireListeners(GANTT_EVENT_TYPE.CLICK_DEPENDENCY_LINK_POINT, {
+          event: e.nativeEvent,
+          index: taskIndex,
+          point: isClickLeftLinkPoint ? 'start' : 'end',
+          record
+        });
+      }
+      stateManager.hideTaskBarSelectedBorder();
     } else if (isClickLeftLinkPoint && event.poniterState === 'draging') {
       if (stateManager.isCreatingDependencyLine()) {
         const link = stateManager.endCreateDependencyLine(e.offset.y);
         if (gantt.hasListeners(GANTT_EVENT_TYPE.CREATE_DEPENDENCY_LINK)) {
           gantt.fireListeners(GANTT_EVENT_TYPE.CREATE_DEPENDENCY_LINK, {
+            federatedEvent: e,
             event: e.nativeEvent,
             link
           });
@@ -361,6 +378,7 @@ function bindTableGroupListener(event: EventManager) {
         const link = stateManager.endCreateDependencyLine(e.offset.y);
         if (gantt.hasListeners(GANTT_EVENT_TYPE.CREATE_DEPENDENCY_LINK)) {
           gantt.fireListeners(GANTT_EVENT_TYPE.CREATE_DEPENDENCY_LINK, {
+            federatedEvent: e,
             event: e.nativeEvent,
             link
           });