+++ packages/vtable/src/PivotTable.ts
@@ -220,11 +220,11 @@ export class PivotTable extends BaseTable implements PivotTableAPI {
       }
       if (this.options.emptyTip) {
         if (this.internalProps.emptyTip) {
-          this.internalProps.emptyTip.resetVisible();
+          this.internalProps.emptyTip?.resetVisible();
         } else {
           const EmptyTip = Factory.getComponent('emptyTip') as IEmptyTipComponent;
           this.internalProps.emptyTip = new EmptyTip(this.options.emptyTip, this);
-          this.internalProps.emptyTip.resetVisible();
+          this.internalProps.emptyTip?.resetVisible();
         }
       }
       //为了确保用户监听得到这个事件 这里做了异步 确保vtable实例已经初始化完成
@@ -436,11 +436,11 @@ export class PivotTable extends BaseTable implements PivotTableAPI {
     }
     if (this.options.emptyTip) {
       if (this.internalProps.emptyTip) {
-        this.internalProps.emptyTip.resetVisible();
+        this.internalProps.emptyTip?.resetVisible();
       } else {
         const EmptyTip = Factory.getComponent('emptyTip') as IEmptyTipComponent;
         this.internalProps.emptyTip = new EmptyTip(this.options.emptyTip, this);
-        this.internalProps.emptyTip.resetVisible();
+        this.internalProps.emptyTip?.resetVisible();
       }
     }
     // this.render();
@@ -1112,9 +1112,31 @@ export class PivotTable extends BaseTable implements PivotTableAPI {
       const dimensions = item.dimensions;
       const width = item.width;
       const cell = this.getCellAddressByHeaderPaths(dimensions);
-      if (cell && !this.internalProps._widthResizedColMap.has(cell.col)) {
-        this._setColWidth(cell.col, width);
-        this.internalProps._widthResizedColMap.add(cell.col); // add resize tag
+      if (cell && cell.col >= this.rowHeaderLevelCount) {
+        const cellPath = this.getCellHeaderPaths(cell.col, this.columnHeaderLevelCount); //如单指标隐藏指标情况，从body行去取headerPath才会包括指标维度
+        if (cellPath.colHeaderPaths.length === dimensions.length) {
+          let match = true;
+          for (let i = 0; i < dimensions.length; i++) {
+            const dimension = dimensions[i];
+            const finded = (cellPath.colHeaderPaths as IDimensionInfo[]).findIndex((colPath: IDimensionInfo, index) => {
+              if (colPath.indicatorKey === dimension.indicatorKey) {
+                return true;
+              }
+              if (colPath.dimensionKey === dimension.dimensionKey && colPath.value === dimension.value) {
+                return true;
+              }
+              return false;
+            });
+            if (finded < 0) {
+              match = false;
+              break;
+            }
+          }
+          if (match && !this.internalProps._widthResizedColMap.has(cell.col)) {
+            this._setColWidth(cell.col, width);
+            this.internalProps._widthResizedColMap.add(cell.col); // add resize tag
+          }
+        }
       }
     }
   }
@@ -1592,11 +1614,11 @@ export class PivotTable extends BaseTable implements PivotTableAPI {
     this.eventManager.updateEventBinder();
     if (this.options.emptyTip) {
       if (this.internalProps.emptyTip) {
-        this.internalProps.emptyTip.resetVisible();
+        this.internalProps.emptyTip?.resetVisible();
       } else {
         const EmptyTip = Factory.getComponent('emptyTip') as IEmptyTipComponent;
         this.internalProps.emptyTip = new EmptyTip(this.options.emptyTip, this);
-        this.internalProps.emptyTip.resetVisible();
+        this.internalProps.emptyTip?.resetVisible();
       }
     }
   }