+++ packages/vtable/src/core/BaseTable.ts
@@ -126,7 +126,6 @@ import { NumberRangeMap } from '../layout/row-height-map';
 import { ListTable } from '../ListTable';
 import type { SimpleHeaderLayoutMap } from '../layout';
 import { RowSeriesNumberHelper } from './row-series-number-helper';
-import { CustomCellStylePlugin, mergeStyle } from '../plugins/custom-cell-style';
 import { hideCellSelectBorder, restoreCellSelectBorder } from '../scenegraph/select/update-select-border';
 import type { ITextGraphicAttribute } from '@src/vrender';
 import { ReactCustomLayout } from '../components/react/react-custom-layout';
@@ -150,6 +149,7 @@ import { setIconColor } from '../icons';
 import { TableAnimationManager } from './animation';
 import type { ITableAnimationOption } from '../ts-types/animation/appear';
 import { checkCellInSelect } from '../state/common/check-in-select';
+import type { CustomCellStylePlugin, ICustomCellStylePlugin } from '../plugins/custom-cell-style';
 
 const { toBoxArray } = utilStyle;
 const { isTouchEvent } = event;
@@ -218,7 +218,7 @@ export abstract class BaseTable extends EventTarget implements BaseTableAPI {
   isReleased: boolean = false;
   _chartEventMap: Record<string, { query?: any; callback: AnyFunction }[]> = {};
 
-  customCellStylePlugin: CustomCellStylePlugin;
+  customCellStylePlugin?: CustomCellStylePlugin;
 
   columnWidthComputeMode?: 'normal' | 'only-header' | 'only-body';
 
@@ -499,11 +499,14 @@ export abstract class BaseTable extends EventTarget implements BaseTableAPI {
 
     internalProps.customMergeCell = options.customMergeCell;
 
-    this.customCellStylePlugin = new CustomCellStylePlugin(
-      this,
-      options.customCellStyle ?? [],
-      options.customCellStyleArrangement ?? []
-    );
+    const CustomCellStylePlugin = Factory.getComponent('customCellStylePlugin') as ICustomCellStylePlugin;
+    if (CustomCellStylePlugin) {
+      this.customCellStylePlugin = new CustomCellStylePlugin(
+        this,
+        options.customCellStyle ?? [],
+        options.customCellStyleArrangement ?? []
+      );
+    }
   }
   /** 节流绘制 */
   throttleInvalidate = throttle2(this.render.bind(this), 200);
@@ -2412,7 +2415,7 @@ export abstract class BaseTable extends EventTarget implements BaseTableAPI {
 
     internalProps.customMergeCell = options.customMergeCell;
 
-    this.customCellStylePlugin.updateCustomCell(
+    this.customCellStylePlugin?.updateCustomCell(
       options.customCellStyle ?? [],
       options.customCellStyleArrangement ?? []
     );
@@ -3293,6 +3296,8 @@ export abstract class BaseTable extends EventTarget implements BaseTableAPI {
    * @returns
    */
   _canDragHeaderPosition(col: number, row: number): boolean {
+    const disableSelect = this.options.select?.disableSelect;
+    const cellDisable = typeof disableSelect === 'function' ? disableSelect(col, row, this) : disableSelect;
     if (
       this.isHeader(col, row) &&
       (this.stateManager.isSelected(col, row) ||
@@ -3301,7 +3306,7 @@ export abstract class BaseTable extends EventTarget implements BaseTableAPI {
             this.getCellRange(this.stateManager.select.cellPos.col, this.stateManager.select.cellPos.row)
           ])) ||
         this.options.select?.disableHeaderSelect ||
-        this.options.select?.disableSelect)
+        cellDisable)
     ) {
       if (this.internalProps.frozenColDragHeaderMode === 'disabled' && this.isFrozenColumn(col)) {
         return false;
@@ -4110,14 +4115,14 @@ export abstract class BaseTable extends EventTarget implements BaseTableAPI {
     //
   }
   hasCustomCellStyle(customStyleId: string): boolean {
-    return this.customCellStylePlugin.hasCustomCellStyle(customStyleId);
+    return this.customCellStylePlugin?.hasCustomCellStyle(customStyleId);
   }
   registerCustomCellStyle(customStyleId: string, customStyle: ColumnStyleOption | undefined | null) {
-    this.customCellStylePlugin.registerCustomCellStyle(customStyleId, customStyle);
+    this.customCellStylePlugin?.registerCustomCellStyle(customStyleId, customStyle);
   }
 
   arrangeCustomCellStyle(cellPos: { col?: number; row?: number; range?: CellRange }, customStyleId: string) {
-    this.customCellStylePlugin.arrangeCustomCellStyle(cellPos, customStyleId);
+    this.customCellStylePlugin?.arrangeCustomCellStyle(cellPos, customStyleId);
   }
   isSeriesNumber(col: number, row: number): boolean {
     return this.internalProps.layoutMap.isSeriesNumber(col, row);
@@ -4159,9 +4164,9 @@ export abstract class BaseTable extends EventTarget implements BaseTableAPI {
   //   startInertia(0, -1, 1, this.stateManager);
   // }
 
-  checkReactCustomLayout(removeAllContainer: () => void) {
+  checkReactCustomLayout() {
     if (!this.reactCustomLayout) {
-      this.reactCustomLayout = new ReactCustomLayout(removeAllContainer, this);
+      this.reactCustomLayout = new ReactCustomLayout(this);
     }
   }