+++ packages/vtable/CHANGELOG.md
@@ -1,6 +1,41 @@
 # Change Log - @visactor/vtable
 
-This log was last generated on Tue, 26 Nov 2024 09:13:29 GMT and should not be manually modified.
+This log was last generated on Fri, 29 Nov 2024 07:59:16 GMT and should not be manually modified.
+
+## 1.11.5
+Fri, 29 Nov 2024 07:59:16 GMT
+
+### Updates
+
+- docs: 更新进入或离开节点时的事件文档
+
+
+- docs: 更新甘特图事件文档，暴露整个e的信息
+
+
+- feat: add @visactor/vtable-plugins package
+
+## 1.11.4
+Fri, 29 Nov 2024 03:38:27 GMT
+
+### Updates
+
+- fix: columnWidthConfig match problem #2957
+
+
+
+## 1.11.3
+Thu, 28 Nov 2024 12:26:11 GMT
+
+### Updates
+
+- fix: frozen chart not update when resize row or column width #2876
+
+
+- refactor: when columnWidthConfig set wrong dimensions should judge this case #2948
+
+
+- fix: fix custom-component update in deleteRecords api
 
 ## 1.11.2
 Tue, 26 Nov 2024 09:13:29 GMT